rideshare-backend/ride-service/src/requirements.txt
Django==5.1.5
djangorestframework==3.14.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.6
pytest==7.4.4
pytest-django==4.7.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
pika==1.3.1
requests==2.31.0
